+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [ 1.1.1 ] - 2026-09-26
+### Changed
+- Changed the default canonical maximum `validTill` from `9999-12-31T00:00:00Z` to `9999-12-31T12:00:00Z`.
 
 ## [ 1.1.0 ] - 2026-09-26
 ### Added